requirements/base.txt
amqp==5.3.1
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.5
cron-descriptor==1.4.5
cryptography==46.0.7
daphne==4.2.1
dj-database-url==3.1.2
Django==6.0.4
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
redis==7.4.0
pytest-cov==7.1.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
semantic-version==2.10.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.3
